Move and consolidate tests
