+++ launch/test/launch/launch_description_source/test_python_launch_description_source.py
@@ -17,7 +17,6 @@
 import os
 
 from launch import LaunchContext
-from launch.actions import IncludeLaunchDescription
 from launch.launch_description_sources import InvalidPythonLaunchFileError
 from launch.launch_description_sources import PythonLaunchDescriptionSource
 
@@ -42,39 +41,3 @@ def test_python_launch_description_source():
     with pytest.raises(FileNotFoundError):
         plds = PythonLaunchDescriptionSource('does_not_exist')
         ld = plds.get_launch_description(LaunchContext())
-
-
-def test_include_python():
-    """Test PythonLaunchDescriptionSource with IncludeLaunchDescription action."""
-    this_dir = os.path.dirname(os.path.abspath(__file__))
-    simple_launch_file_path = os.path.join(this_dir, 'simple.launch.py')
-    plds = PythonLaunchDescriptionSource(simple_launch_file_path)
-
-    action = IncludeLaunchDescription(plds)
-    assert 'IncludeLaunchDescription' in action.describe()
-    assert isinstance(action.describe_sub_entities(), list)
-    assert isinstance(action.describe_conditional_sub_entities(), list)
-    # Result should only contain a single launch description as there are no launch arguments.
-    assert len(action.visit(LaunchContext())) == 1
-    assert action.get_asyncio_future() is None
-    assert len(action.launch_arguments) == 0
-
-    assert action.launch_description_source.location == simple_launch_file_path
-
-
-def test_include_python_implicit():
-    """Test IncludeLaunchDescription with implicit inclusion of PythonLaunchDescriptionSource."""
-    this_dir = os.path.dirname(os.path.abspath(__file__))
-    simple_launch_file_path = os.path.join(this_dir, 'simple.launch.py')
-
-    # pass in string instead of PythonLaunchDescriptionSource
-    action = IncludeLaunchDescription(simple_launch_file_path)
-    assert 'IncludeLaunchDescription' in action.describe()
-    assert isinstance(action.describe_sub_entities(), list)
-    assert isinstance(action.describe_conditional_sub_entities(), list)
-    # Result should only contain a single launch description as there are no launch arguments.
-    assert len(action.visit(LaunchContext())) == 1
-    assert action.get_asyncio_future() is None
-    assert len(action.launch_arguments) == 0
-
-    assert action.launch_description_source.location == simple_launch_file_path
